@@ -20,7 +20,7 @@ repository = "https://github.com/dfinity/agent-rs"
 # MSRV
 # Avoid updating this field unless we use new Rust features
 # Sync rust-version in rust-toolchain.toml
-rust-version = "1.85.0"
+rust-version = "1.88.0"
 license = "Apache-2.0"
 
 [workspace.lints.clippy]
